Refactor DHT22 sensor to individual module.
