+++ src/modules/blockly/generators/propc/sensors.js
@@ -3035,100 +3035,6 @@ Blockly.propc.keypad_read = function() {
   }
 };
 
-
-// ------------------ DHT22 Temp & Humidity Sensor ----------------------
-
-/**
- *
- * @type {{
- *  init: Blockly.Blocks.dht22_read.init,
- *  addPinMenu: *,
- *  mutationToDom: *,
- *  helpUrl: string,
- *  setToOther: *,
- *  domToMutation: *
- *  }}
- */
-Blockly.Blocks.dht22_read = {
-  helpUrl: Blockly.MSG_TEMPERATURE_HELPURL,
-  init: function() {
-    this.setTooltip(Blockly.MSG_DHT22_READ_TOOLTIP);
-    this.setColour(colorPalette.getColor('input'));
-    this.pinChoices = ['PIN'];
-    this.otherPin = [false];
-    this.addPinMenu('Temp & Humidity read PIN', null, 0);
-    this.setPreviousStatement(true, 'Block');
-    this.setInputsInline(true);
-    this.setNextStatement(true, null);
-  },
-  mutationToDom: Blockly.Blocks['sensor_ping'].mutationToDom,
-  domToMutation: Blockly.Blocks['sensor_ping'].domToMutation,
-  addPinMenu: Blockly.Blocks['sensor_ping'].addPinMenu,
-  setToOther: Blockly.Blocks['sensor_ping'].setToOther,
-};
-
-/**
- *
- * @return {string}
- */
-Blockly.propc.dht22_read = function() {
-  if (!this.disabled) {
-    Blockly.propc.definitions_['dht22'] = '#include "dht22.h"';
-  }
-
-  let pin = '0';
-  if (this.otherPin[0]) {
-    pin = Blockly.propc.valueToCode(
-        this, this.pinChoices[0], Blockly.propc.ORDER_ATOMIC) || '0';
-  } else {
-    pin = this.getFieldValue(this.pinChoices[0]);
-  }
-
-  return 'dht22_read(' + pin + ');';
-};
-
-/**
- *
- * @type {{
- *  init: Blockly.Blocks.dht22_value.init,
- *  helpUrl: string
- *  }}
- */
-Blockly.Blocks.dht22_value = {
-  helpUrl: Blockly.MSG_TEMPERATURE_HELPURL,
-  init: function() {
-    this.setTooltip(Blockly.MSG_DHT22_VALUE_TOOLTIP);
-    this.setColour(colorPalette.getColor('input'));
-    this.appendDummyInput()
-        .appendField('Temp & Humidity')
-        .appendField(new Blockly.FieldDropdown([
-          ['temperature (\u00b0F)', 'Temp,FAHRENHEIT'],
-          ['temperature (\u00b0C)', 'Temp,CELSIUS'],
-          ['temperature (Kelvin)', 'Temp,KELVIN'],
-          ['relative humidity (%)', 'Humidity,'],
-        ]), 'ACTION')
-        .appendField('\u2715 10');
-    this.setInputsInline(false);
-    this.setOutput(true, 'Number');
-  },
-};
-
-/**
- *
- * @return {[string, number]}
- */
-Blockly.propc.dht22_value = function() {
-  if (!this.disabled) {
-    Blockly.propc.definitions_['dht22'] = '#include "dht22.h"';
-  }
-  const action = this.getFieldValue('ACTION').split(',');
-  return [
-    'dht22_get' + action[0] + '(' + action[1] + ')',
-    Blockly.propc.ORDER_ATOMIC,
-  ];
-};
-
-
 // ------------------ BME680 Air Quality Sensor -----------------------------
 
 /**
